Changes in Valhalla-Master:
- Building as submodule fixes (3782)
- add distance limit in error message for 154 error code (3779)
- fix invalid speed detection in GraphTile::GetSpeed (3800)
- osmway struct update:  added up to 33 and not 32 (3808)
- prep for 3.2.0 (3809)
- Update api-reference.md
- valhalla_run_route was missing config logic. (3824)
- Dev/tag highway ferry (3816)
- Add time tracking to matrix (3795)
- add 'available_actions' to /status response (3836)
- Fix mising single quote at the end of sample command (3841)
- Handle hexlifying strings with unsigned chars (3842)
- Use snprintf instead of sprintf (3847)
- Build all of Mjolnir with -Werror (3845)
- initialize destinations only once in timedistance matrix (3830)
- add "waiting time" to locations (3849)
- add more options to valhalla_build_extract (3856)
- Integers to expansion JSON output were cast wrongly (3857)
- fix hazmat (3865)
- admin builder with c geos (3683)
- make as many modes accessible as possible in the reclassification logic (3860)
- fix valhalla_build_extract to derive tile id & level from the relative tile path, not the full one (3864)
- fix valhalla_ways_to_edges (3872)
- GraphReader with write-access to traffic extract (3876)

Related work items: #105004

+++ valhalla/baldr/graphtile.h
@@ -627,7 +627,7 @@ public:
    */
   inline uint32_t GetSpeed(const DirectedEdge* de,
                            uint8_t flow_mask = kConstrainedFlowMask,
-                           uint32_t seconds = kInvalidSecondsOfWeek,
+                           uint64_t seconds = kInvalidSecondsOfWeek,
                            bool is_truck = false,
                            uint8_t* flow_sources = nullptr,
                            const uint64_t seconds_from_now = 0) const {
@@ -705,6 +705,7 @@ public:
 
     // fallback to constrained if time of week is within 7am to 7pm (or if no time was passed in) and
     // if the edge has constrained speed
+    // kInvalidSecondsOfWeek %= midgard::kSecondsPerDay = 12.1
     seconds %= midgard::kSecondsPerDay;
     auto is_daytime = (25200 < seconds && seconds < 68400);
     if ((invalid_time || is_daytime) && (flow_mask & kConstrainedFlowMask) &&
